check uniqueness of member_config value while member/configComplete action (fixes #4018, BP from #4012)
(cherry picked from commit 4769d307b49895abfdbe749b734d105819bdf9d4)

+++ lib/action/opMemberAction.class.php
@@ -309,6 +309,13 @@ abstract class opMemberAction extends sfActions
         }
         $config->setValue($pre->getValue());
 
+        if (!$config->checkUniqueness())
+        {
+          $this->getUser()->setFlash('error', 'The inputted value is already exist.');
+
+          return sfView::SUCCESS;
+        }
+
         $config->save();
 
         $pre->delete();
